requirements.txt
alembic
astroid
bitsets
Click
concepts
dnspython
fileconfig
Flask
Flask-Migrate
Flask-SQLAlchemy
gitdb2
GitPython
graphviz
greenlet
gunicorn
isort
ItsDangerous
Jinja2
lazy-object-proxy
Mako
MarkupSafe
mccabe
monotonic
mysqlclient
pylint
python-dateutil
python-editor
python-engineio
python-socketio
six
smmap2
SQLAlchemy
Werkzeug
wrapt==1.10.11
